🎨 Move fetching rewards to week model
